+++ providers/ic/src/types/publish.metas.ts
@@ -1,4 +1,4 @@
-import { Meta } from "@deckdeckgo/editor";
+import {Meta} from '@deckdeckgo/editor';
 
 export interface PublishMeta {
   meta: Meta;