+++ providers/ic/src/utils/publish.resources.utils.ts
@@ -35,7 +35,9 @@ export const uploadResources = async ({meta}: {meta: Meta | undefined}) => {
 
   // 4. We only upload resources that have not been yet uploaded. In other words: we upload the resources the first time or if hashes are modified.
   // TODO: remove temporary rollout of last CSS update
-  const kitNewFiles: Kit[] = kit.filter(({filename, mimeType}: Kit) => !keys.includes(filename) || mimeType === 'text/css');
+  const kitNewFiles: Kit[] = kit.filter(
+    ({filename, mimeType}: Kit) => !keys.includes(filename) || mimeType === 'text/css'
+  );
 
   if (!kitNewFiles || kitNewFiles.length <= 0) {
     return;