+++ providers/ic/CHANGELOG.md
@@ -66,7 +66,7 @@
 ### Fix
 
 - do not update `published_at` date if already published - keep original publish date
-- remove unused dtd in rss and move `items` as `channel` child 
+- remove unused dtd in rss and move `items` as `channel` child
 
 # 0.0.9 (2022-05-22)